data-table: rename model module to table
