+++ packages/data-table/test/adapter-integration-contract.ts
@@ -3,7 +3,7 @@ import { beforeEach, it } from 'node:test'
 import { boolean, nullable, number, string } from '@remix-run/data-schema'
 
 import type { Database } from '../src/lib/database.ts'
-import { createTable } from '../src/lib/model.ts'
+import { createTable } from '../src/lib/table.ts'
 import { between, eq, ilike, inList, ne } from '../src/lib/operators.ts'
 
 let Accounts = createTable({
